requirements.txt
absl-py                      1.0.0
affine                       2.3.0
argcomplete                  1.12.3
argon2-cffi                  20.1.0
astunparse                   1.6.3
async-generator              1.10
attrs                        21.2.0
backcall                     0.2.0
bleach                       4.0.0
brotlipy                     0.7.0
cached-property              1.5.2
cachetools                   4.2.4
certifi                      2021.5.30
cffi                         1.15.0
chardet                      4.0.0
charset-normalizer           2.0.8
click                        8.0.3
click-plugins                1.1.1
cligj                        0.7.2
colorama                     0.4.4
cryptography                 35.0.0
cycler                       0.10.0
debugpy                      1.5.1
decorator                    5.1.0
defusedxml                   0.7.1
entrypoints                  0.3
et-xmlfile                   1.1.0
Flask                        2.0.2
flatbuffers                  2.0
fonttools                    4.28.2
gast                         0.4.0
GDAL                         3.0.2
google-auth                  2.3.3
google-auth-oauthlib         0.4.6
google-pasta                 0.2.0
grpcio                       1.42.0
h5py                         3.6.0
idna                         2.10
importlib-metadata           4.8.2
ipykernel                    6.4.1
ipython                      7.29.0
ipython-genutils             0.2.0
itsdangerous                 2.0.1
jedi                         0.18.0
Jinja2                       3.0.3
joblib                       1.1.0
jsonschema                   3.2.0
jupyter-client               7.0.6
jupyter-core                 4.9.1
jupyterlab-pygments          0.1.2
keras                        2.7.0
Keras-Preprocessing          1.1.2
kiwisolver                   1.3.1
libclang                     12.0.0
llvmlite                     0.35.0
Markdown                     3.3.6
MarkupSafe                   2.0.1
matplotlib                   3.5.0
matplotlib-inline            0.1.2
mistune                      0.8.4
mkl-fft                      1.3.0
mkl-random                   1.1.1
mkl-service                  2.3.0
nbclient                     0.5.3
nbconvert                    6.1.0
nbformat                     5.1.3
nest-asyncio                 1.5.1
notebook                     6.4.6
numba                        0.52.0
numpy                        1.19.2
oauthlib                     3.1.1
olefile                      0.46
opencv-python                4.5.4.60
openpyxl                     3.0.9
opt-einsum                   3.3.0
packaging                    21.3
pandas                       1.3.4
pandocfilters                1.4.3
parso                        0.8.2
pickleshare                  0.7.5
Pillow                       8.4.0
pip                          21.2.4
prometheus-client            0.12.0
prompt-toolkit               3.0.20
protobuf                     3.19.1
pyasn1                       0.4.8
pyasn1-modules               0.2.8
pycparser                    2.21
pydub                        0.25.1
Pygments                     2.10.0
pyOpenSSL                    21.0.0
pyparsing                    2.4.7
pyrsistent                   0.18.0
PySocks                      1.7.1
python-dateutil              2.8.2
python-dotenv                0.19.2
pytz                         2021.3
pywin32                      228
pywinpty                     0.5.7
PyYAML                       6.0
pyzmq                        22.3.0
rasterio                     1.1.0
requests                     2.26.0
requests-oauthlib            1.3.0
roboflow                     0.2.1
rsa                          4.8
scikit-learn                 1.0.1
scipy                        1.7.3
seaborn                      0.11.2
selenium                     3.141.0
Send2Trash                   1.8.0
setuptools                   58.0.4
setuptools-scm               6.3.2
six                          1.16.0
sklearn                      0.0
snuggs                       1.4.7
tensorboard                  2.7.0
tensorboard-data-server      0.6.1
tensorboard-plugin-wit       1.8.0
tensorflow                   2.7.0
tensorflow-estimator         2.7.0
tensorflow-io-gcs-filesystem 0.22.0
termcolor                    1.1.0
terminado                    0.9.4
testpath                     0.5.0
thop                         0.0.31.post2005241907
threadpoolctl                3.0.0
tomli                        1.2.2
torch                        1.9.0+cu111
torchaudio                   0.9.0
torchvision                  0.10.0+cu111
tornado                      6.1
tqdm                         4.62.3
traitlets                    5.1.1
typing_extensions            4.0.1
urllib3                      1.26.6
utils                        1.0.1
wcwidth                      0.2.5
webencodings                 0.5.1
Werkzeug                     2.0.2
wget                         3.2
wheel                        0.37.0
win-inet-pton                1.1.0
wincertstore                 0.2
wrapt                        1.13.3
zipp                         3.6.0